cmd - echo, export, unset, pwd, printenv
