+++ develop/libs/libft/ft_putendl_fd.c
@@ -14,6 +14,7 @@
 
 void	ft_putendl_fd(char *s, int fd)
 {
-	ft_putstr_fd(s, fd);
+	if (s)
+		ft_putstr_fd(s, fd);
 	ft_putchar_fd('\n', fd);
 }
