+++ develop/srcs/main.c
@@ -112,7 +112,6 @@ int	main(int argc, char **argv, char **envp)
 		out = readline("minishell $ ");
 		if (!check_args(out, &prompt))
 			break ;
-		process(&prompt);
 	}
 	exit (exit_status);
 }
