+++ develop/srcs/utils/matrix.c
@@ -1,0 +1,36 @@
+# include "minishell.h"
+
+int	get_matrixlen(char **m)
+{
+	int	i;
+
+	i = 0;
+	while (m && m[i])
+		i++;
+	return (i);
+}
+
+char	**dup_matrix(char **m)
+{
+	char	**out;
+	int		n_rows;
+	int		i;
+
+	i = 0;
+	n_rows = get_matrixlen(m);
+	out = malloc(sizeof(char *) * (n_rows + 1));
+	if (!out)
+		return (NULL);
+	while (m[i])
+	{
+		out[i] = ft_strdup(m[i]);
+		if (!out[i])
+		{
+			free_matrix(&out);
+			return (NULL);
+		}
+		i++;
+	}
+	out[i] = NULL;
+	return (out);
+}
