+++ develop/srcs/utils/utils.c
@@ -1,0 +1,49 @@
+#include "minishell.h"
+
+int	ft_strchr_int(const char *s, int c)
+{
+	unsigned char	c_unsigned;
+	int				i;
+
+	i = 0;
+	if (!s)
+		return (-1);
+	c_unsigned = (unsigned char)c;
+	while (s[i] != '\0')
+	{
+		if (s[i] == c_unsigned)
+			return (i);
+		i++;
+	}
+	if (c_unsigned == '\0')
+		return (i);
+	return (-1);
+}
+
+int	token_countcmd(t_token *token)
+{
+	int	ans;
+
+	ans = 0;
+	while (token)
+	{
+		if (token->type == 1)
+			ans += 1;
+		token = token->next;
+	}
+	return (ans);
+}
+
+void	ft_lstclear(t_list **lst, void (*del)(void *))
+{
+	t_list	*tmp;
+
+	if (!del)
+		return ;
+	while (*lst)
+	{
+		tmp = (*lst)->next;
+		ft_lstdelone(*lst, del);
+		(*lst) = tmp;
+	}
+}
